requirements2.txt
Flask==2.1.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
pymongo[srv]==3.12.1
pyyaml==5.4.1
Werkzeug==2.0.2